requirements/base.txt
Django==5.0.2
django-environ
psycopg2-binary
djangorestframework

celery
django-celery-results
django-celery-beat

whitenoise

django-filter
django-extensions
django-cors-headers
django-storages

drf-jwt

boto3
attrs

djangorestframework-simplejwt
drf-spectacular

django-redis
Faker
factory-boy
pytest
pytest-django
pillow
elasticsearch-dsl
django-elasticsearch-dsl
az-iranian-bank-gateways